Added a new Build with limit on treats
Ok, added a limit to how many treats you can through. It's a public
variable so we can easily change that during the "Juicing" period. I'm
also going to make this an entirely new build, since the demo we had was
pretty much fine.

I'll begin working on a countdown for treat throwing, maybe.